requirements.txt
# requirements.txt - Updated for PyPSA-Ethiopia

# Core Streamlit and web framework
streamlit>=1.28.0
streamlit-folium>=0.15.0

# Data handling and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# PyPSA and energy system modeling
pypsa>=0.26.0
# Note: PyPSA will be optional for the data download functionality

# Geospatial and GIS
geopandas>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
rasterio>=1.3.0

# Data processing
xarray>=2023.0.0
netcdf4>=1.6.0
bottleneck>=1.3.0
numexpr>=2.8.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
folium>=0.14.0

# HTTP requests and downloads
requests>=2.31.0
urllib3>=2.0.0
tqdm>=4.65.0

# File processing
pyyaml>=6.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Optional solver (recommended)
# highspy>=1.5.0  # Uncomment if available

# Development and utilities
python-dotenv>=1.0.0
pathlib2>=2.3.0  # For Python < 3.4 compatibility
typing-extensions>=4.0.0

# Optional: Database support
# psycopg2-binary>=2.9.0  # For PostgreSQL
# redis>=4.0.0  # For caching

# Google Drive downloads (used by data downloader)
googledrivedownloader>=0.4

# Alternative download methods
zipfile38>=0.1.0; python_version < "3.8"

# Logging and monitoring
colorlog>=6.7.0